+++ CHANGELOG.md
@@ -12,11 +12,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add extra constructors on Money to create with a given MoneyContext, instead of MidpointRounding.
 - Add MoneyContext property on Money.
 - Add Scale property on Money.
+- Add Precision property on Money (= still internal!).
 - Add NodaMoney.DependencyInjection package to register MoneyContext .NET DI.
 
 ### Changed
 - Optimized Money struct size. This was 18 bytes (padded 24 bytes), but is now 16 bytes (padded 16 bytes) (33% less).
-  This means Money is the same size as Decimal struct!
+  This means Money is the same size as Decimal struct!!!
 
 ### Removed
 -